TOOLS-4263 Move the throwaway mongod helper into sharedsuite
The dump server-death test starts a mongod of its own so that it can kill it
without disturbing the deployment the rest of the suite shares. A mongoimport test
converted next needs the same thing for a different reason: a replica set that was
never initiated has no primary, which is what that test is about. Rather than stand
up a second copy in `integration/exportimport`, the machinery moves to the package
both suites already embed.

- `integration/dumprestore/server_death_test.go` -> `integration/sharedsuite/throwaway_mongod.go` - `throwawayMongod` and its `startThrowawayMongod`, along with the `LockedBuffer` that collects the process's stderr for a failure message, become exported members of `sharedsuite`. `StartThrowawayMongod` now takes extra mongod arguments, so a caller can ask for one that belongs to a named replica set.

No behavior changes. The server-death test's assertions are untouched; only the
names it calls them through move.
